requirements.txt
openai>=1.59.3
python-dotenv>=1.0.1
python-docx>=1.1.2
markdown2>=2.5.2
weasyprint>=63.1
pdfminer.six>=20240706
gradio>=5.9.1
jinja2>=3.1.5 